+++ test/connect/KafkaConnection.test.ts
@@ -27,7 +27,7 @@ suite('KafkaConnection', ()=> {
             'credential.password', brokerPass,
             'credential.mechanism', 'plain',
             'options.num_partitions', 2,
-            'options.readable_partitions', '1',
+            'options.read_partitions', '1',
             'options.write_partition', '1'
         );