+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Kafka components for Pip.Services in Node.js / ES2017 Changelog
 
+## <a name="1.1.0"></a> 1.1.0 (2022-09-18)
+
+### Features
+* Added **KafkaConnectionListener** component, listen connection in background
+* Added new config params for **KafkaMessageQueue** and **KafkaConnection**
+
 ## <a name="1.0.0"></a> 1.0.0 (2021-05-28)
 
 Initial release